display: Allow HBM in manual brightness if there is no time limitation
If a display does not set time limitation for HBM, it means it can
safely run at HBM for as long as it wants. With this in mind, we can
allow HBM in manual brightness too for such displays. This still obeys
the thermal throttling and battery saver conditions for obvious reasons.

Change-Id: I18ca7748c1f32dca1e70497afd0b6df8ed81ad5d